[cogs/fun/rng.py] Split off rng_fun.py for higher granularity
