+++ cogs/fun/rng.py
@@ -1,9 +1,8 @@
 #!/usr/bin/env python3
 # pylint: disable=C0103
 
-"""Commands that invoke random things. Coin flips, dice rolls, kemonomimi... wait."""
+"""Commands that invoke random things, part 1."""
 
-import json
 import random
 import re
 
@@ -18,11 +17,6 @@ MAX_ROLL_SIZE = 30
 MAX_DIE_SIZE = 2000
 
 URL_RANDOM_WORD_API = "http://setgetgo.com/randomword/get.php"
-URL_RANDOM_DOG_API = "https://random.dog/woof.json"
-URL_RANDOM_CAT_API = "https://random.cat/meow"
-URL_RANDOM_BIRB = "https://random.birb.pw/img/{0}"
-URL_RANDOM_BIRB_API = "https://random.birb.pw/tweet.json/"
-URL_RANDOM_NEKO_API = "https://nekos.life/api/neko"
 
 systemrandom = random.SystemRandom()
 
@@ -71,58 +65,6 @@ def parse_rolls(*expressions, **kwargs):
 
 class Random:
     """Commands that produce random outputs."""
-
-    @commands.command(aliases=["doge"])
-    @commands.cooldown(6, 12)
-    async def dog(self, ctx):
-        """Fetch a random dog."""
-        async with ctx.bot.session.get(URL_RANDOM_DOG_API) as response:
-            if response.status == 200:
-                data = await response.text()
-                doggo = json.loads(data)
-                url = doggo["url"]
-                await ctx.send(url)
-            else:
-                await ctx.send("Could not reach random.dog. :<")
-
-    @commands.command(aliases=["feline"])
-    @commands.cooldown(6, 12)
-    async def cat(self, ctx):
-        """Fetch a random cat."""
-        async with ctx.bot.session.get(URL_RANDOM_CAT_API) as response:
-            if response.status == 200:
-                data = await response.text()
-                catto = json.loads(data)
-                url = catto["file"]
-                await ctx.send(url)
-            else:
-                await ctx.send("Could not reach random.cat. :<")
-
-    @commands.command(aliases=["kemonomimi", "catgirl", "kneko", "nekomimi",
-                               "foxgirl" "kitsune", "kitsunemimi"])
-    @commands.cooldown(6, 12)
-    async def kemono(self, ctx):
-        """Fetch a random animal-eared person."""
-        async with ctx.bot.session.get(URL_RANDOM_NEKO_API) as response:
-            if response.status == 200:
-                neko = await response.json()
-                url = neko["neko"]
-                await ctx.send(url)
-            else:
-                await ctx.send("Could not reach nekos.life. :<")
-
-    @commands.command()
-    @commands.cooldown(6, 12)
-    async def birb(self, ctx):
-        """Fetch a random birb."""
-        async with ctx.bot.session.get(URL_RANDOM_BIRB_API) as response:
-            if response.status == 200:
-                data = await response.text()
-                borb = json.loads(data)
-                url = URL_RANDOM_BIRB.format(borb["file"])
-                await ctx.send(url)
-            else:
-                await ctx.send("Could not reach random.birb.pw. :<")
 
     @commands.command(aliases=["cflip", "coinflip"])
     @commands.cooldown(6, 12)
